// src/api/models/domain_PublicTextBoxesPageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { domain_PublicTextBoxSummary } from './domain_PublicTextBoxSummary';
export type domain_PublicTextBoxesPageResponse = {
    items?: Array<domain_PublicTextBoxSummary>;
    limit?: number;
    offset?: number;
    total?: number;
};