// src/api/models/domain_TextBoxSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { domain_TextBoxStats } from './domain_TextBoxStats';
import type { domain_TextBoxVisibility } from './domain_TextBoxVisibility';
export type domain_TextBoxSummary = {
    createdAt?: string;
    id?: string;
    isDefault?: boolean;
    stats?: domain_TextBoxStats;
    telegramId?: string;
    title?: string;
    updatedAt?: string;
    visibility?: domain_TextBoxVisibility;
};